discord-bot/requirements.txt
python-dotenv 
discord.py==2.4.0
audioop-lts==0.2.1
aiohttp==3.11.12